// src/_requests/Maja/models/NursesQueryPermissionsResponseDataItem.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { NursesQueryPermissionsResponseDataItemPermission } from './NursesQueryPermissionsResponseDataItemPermission';

export type NursesQueryPermissionsResponseDataItem = {
    expire_date?: string;
    id?: number;
    nurseId?: number;
    permission?: NursesQueryPermissionsResponseDataItemPermission;
    permissionId?: number;
};